+++ src/main/java/kr/lnsc/api/link/exception/InvalidUrlException.java
@@ -1,4 +1,4 @@
-package kr.lnsc.api.domain.link.exception;
+package kr.lnsc.api.link.exception;
 
 import kr.lnsc.api.error.ErrorCode;
 import kr.lnsc.api.error.exception.InvalidValueException;
@@ -7,6 +7,6 @@ public class InvalidUrlException extends InvalidValueException {
     private static final String MESSAGE = "유효하지 않는 URL입니다. 확인 후 다시 입력해주세요.";
 
     public InvalidUrlException() {
-        super(MESSAGE, ErrorCode.UNHANDLED_ERROR);
+        super(MESSAGE, ErrorCode.INVALID_URL);
     }
 }
